+++ pages/home/index.js
@@ -7,7 +7,7 @@ import { Sections } from '@components/sections'
 import { Press } from '@components/press'
 import { Videos } from '@components/videos'
 import { CaseStudies } from '@components/case-studies'
-import { meta, caseStudies, videos, press } from './data'
+import { meta, caseStudies, videos, press } from '../../common/data/home'
 import fetch from 'cross-fetch'
 import { News } from '@components/articles'
 import { PhotoGrid } from '@components/photos-grid'
@@ -15,7 +15,7 @@ import { AuthGraphic } from '@components/graphics/auth'
 import { useTrail, animated } from 'react-spring'
 import { useHover } from 'use-events'
 import { transition } from '@common/theme'
-import { apps } from './apps'
+import { apps } from '../../common/data/apps'
 import { AppMiningGraphic } from '@components/graphics/app-mining'
 
 const columns = 7
