--- pages/papers/papers.md
@@ -1,15 +1,0 @@
-####
-
-#### The Blockstack technology has evolved considerably in the past two years and the papers below don’t fully reflect the current decentralized computing stack.
-
-##### Blockstack: A New Decentralized Internet
-
-Whitepaper Version 1.1 - Oct 2017
-
-<a href="https://blockstack.org/whitepaper.pdf" target="_blank" class="button">PDF</a>
-
-##### Blockstack Token Whitepaper
-
-Token Whitepaper Version 1.2 - Oct 2017
-
-<a href="https://blockstack.org/tokenpaper.pdf" target="_blank" class="button">PDF</a>
