+++ pages/technology/index.js
@@ -8,7 +8,7 @@ import { Events } from '@components/events'
 import { Button } from '@components/button'
 import { AppsList } from '@components/apps-list'
 import { PhotoGrid } from '@components/photos-grid'
-import { photos, apps, appBuildersGrid, usersGrid } from './data'
+import { photos, apps, appBuildersGrid, usersGrid } from '../../common/data/technology'
 import { AuthGraphic } from '@components/graphics/auth'
 import { AppMiningGraphic } from '@components/graphics/app-mining'
 import { Codeblock } from '@components/code'
